LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-js-innerhtml.html

<!-- webkit-test-runner [ dumpResourceResponseMIMETypes=* ] -->
<!DOCTYPE html>
<script src="../../../../resources/srcset-helper.js"></script>
<script src="../../../../resources/js-test.js"></script>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
    if (window.internals)
        internals.clearMemoryCache();

    function fileName(src) {
        var array = src.split('/');
        return array[array.length -1];
    };

    function currentSrcFileName(id) {
        var currentSrc = document.getElementById(id).currentSrc;
        return fileName(currentSrc);
    };

    addEventListener("DOMContentLoaded", function() {
        window.testdiv = document.createElement("div");
        window.testdiv.innerHTML = '<img id="foo" src="../resources/blue-100-px-square.png" srcset="../../../hidpi/resources/image-set-1x.png 200w, ../../../hidpi/resources/image-set-2x.png 400w, ../resources/image-set-4x.png 800w">';
    }, false);
    function runTest()
    {
        shouldBe('fileName(window.testdiv.children[0].currentSrc)', '"image-set-4x.png"');
    }
</script>
<img id="foo">
