LayoutTests/platform/win/fast/dom/HTMLImageElement/sizes/image-sizes-js-innerhtml-expected.txt

PASS fileName(window.testdiv.children[0].currentSrc) is "image-set-4x.png"
PASS successfullyParsed is true

TEST COMPLETE

image-sizes-js-innerhtml.html has MIME type text/html
srcset-helper.js has MIME type application/x-javascript
js-test.js has MIME type application/x-javascript
image-set-4x.png has MIME type image/png
image-sizes-js-innerhtml.html has MIME type text/html
srcset-helper.js has MIME type application/x-javascript
js-test.js has MIME type application/x-javascript
image-set-4x.png has MIME type image/png
